fix win build

+++ include/orca/task/GenericTask.h
@@ -52,7 +52,7 @@ public:
 
     virtual ~GenericTask();
 
-    int getWeight() const;
+    double getWeight() const;
 
     void setWeight(double weight);
